Add files via upload
It was conceived as a telegram bot written on telebot, which would automatically send vacancies or orders. I didn’t manage to implement auto-mailing, but I managed to make a subscription (albeit it is paid in words). The bot is connected to a database that stores all user settings and data. This is in general terms, but in general the bot is not completed,
because C# got carried away again.